# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Basketball attendance and player evaluation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team (head_coach only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team (head_coach only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports/team/{id}:
    post:
      operationId: getTeamReport
      tags: [reports]
      summary: Get attendance and evaluation report for a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilter"
      responses:
        "200":
          description: Team report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamReport"
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        ageCategory:
          type: ["string", "null"]
        coachId:
          type: ["integer", "null"]
        coachName:
          type: ["string", "null"]
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        ageCategory:
          type: string
        coachId:
          type: ["integer", "null"]
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        ageCategory:
          type: string
        coachId:
          type: ["integer", "null"]
...
    Criterion:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    PlayerReport:
      type: object
      required: [playerId, playerName, totalSessions, present, absent, late, excused, attendancePct]
      properties:
        playerId:
          type: integer
        playerName:
          type: string
        jerseyNumber:
          type: ["integer", "null"]
        totalSessions:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        excused:
          type: integer
        attendancePct:
          type: number
        overallAvg:
          type: ["number", "null"]
        criteriaAvgs:
          type: array
          items:
            $ref: "#/components/schemas/CriterionAvg"
    CriterionAvg:
      type: object
      required: [criterionId, criterionName]
      properties:
        criterionId:
          type: integer
        criterionName:
          type: string
        avg:
          type: ["number", "null"]
    TeamReport:
      type: object
      required: [teamId, teamName, totalSessions, criteria, players]
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        totalSessions:
          type: integer
        criteria:
          type: array
          items:
            $ref: "#/components/schemas/Criterion"
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerReport"
    ReportFilter:
      type: object
      properties:
        dateFrom:
          type: string
        dateTo:
          type: string
...