# lib/api-zod/src/generated/types/playerReport.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Basketball attendance and player evaluation API
 * OpenAPI spec version: 0.1.0
 */
import type { CriterionAvg } from './criterionAvg';

export interface PlayerReport {
  playerId: number;
  playerName: string;
  /** @nullable */
  jerseyNumber?: number | null;
  totalSessions: number;
  present: number;
  absent: number;
  late: number;
  excused: number;
  attendancePct: number;
  /** @nullable */
  overallAvg?: number | null;
  criteriaAvgs?: CriterionAvg[];
}